refactor(core): decompose graph services
Split graph records, embeddings, entity managers, work items, and
relationships into acyclic canonical modules. Keep services.graph as the
small public contract and migrate private callers to their real owners.

Preserve manager methods, runtime mapping behavior, and exact exception
propagation. Add static export, cycle, and dataclass regressions.
