+++ apps/api/tests/harness/context.py
@@ -9,7 +9,7 @@ from contextlib import asynccontextmanager
 from typing import Any
 from unittest.mock import AsyncMock, patch
 
-from sibyl_core.services import ActiveGraphRuntime
+from sibyl_core.services import GraphRuntime
 from tests.harness.mocks import (
     MockEntityManager,
     MockGraphClient,
@@ -68,8 +68,8 @@ class ToolTestContext:
         async def async_get_graph_client() -> MockGraphClient:
             return self.graph_client
 
-        async def async_get_graph_runtime(group_id: str) -> ActiveGraphRuntime:
-            return ActiveGraphRuntime(
+        async def async_get_graph_runtime(group_id: str) -> GraphRuntime:
+            return GraphRuntime(
                 client=self.graph_client,
                 entity_manager=self.entity_manager,
                 relationship_manager=self.relationship_manager,
@@ -146,7 +146,7 @@ async def mock_graph_connected() -> AsyncGenerator[MockGraphClient]:
         patch(
             "sibyl_core.tools.search.get_graph_runtime",
             AsyncMock(
-                return_value=ActiveGraphRuntime(
+                return_value=GraphRuntime(
                     client=client,
                     entity_manager=MockEntityManager(),
                     relationship_manager=MockRelationshipManager(),
@@ -156,7 +156,7 @@ async def mock_graph_connected() -> AsyncGenerator[MockGraphClient]:
         patch(
             "sibyl_core.tools.explore.get_graph_runtime",
             AsyncMock(
-                return_value=ActiveGraphRuntime(
+                return_value=GraphRuntime(
                     client=client,
                     entity_manager=MockEntityManager(),
                     relationship_manager=MockRelationshipManager(),
@@ -166,7 +166,7 @@ async def mock_graph_connected() -> AsyncGenerator[MockGraphClient]:
         patch(
             "sibyl_core.tools.add.get_graph_runtime",
             AsyncMock(
-                return_value=ActiveGraphRuntime(
+                return_value=GraphRuntime(
                     client=client,
                     entity_manager=MockEntityManager(),
                     relationship_manager=MockRelationshipManager(),
@@ -177,7 +177,7 @@ async def mock_graph_connected() -> AsyncGenerator[MockGraphClient]:
         patch(
             "sibyl_core.tools.manage.get_graph_runtime",
             AsyncMock(
-                return_value=ActiveGraphRuntime(
+                return_value=GraphRuntime(
                     client=client,
                     entity_manager=MockEntityManager(),
                     relationship_manager=MockRelationshipManager(),
@@ -188,7 +188,7 @@ async def mock_graph_connected() -> AsyncGenerator[MockGraphClient]:
         patch(
             "sibyl_core.tools.health.get_graph_runtime",
             AsyncMock(
-                return_value=ActiveGraphRuntime(
+                return_value=GraphRuntime(
                     client=client,
                     entity_manager=MockEntityManager(),
                     relationship_manager=MockRelationshipManager(),
@@ -200,7 +200,7 @@ async def mock_graph_connected() -> AsyncGenerator[MockGraphClient]:
         patch(
             "sibyl_core.tools.admin.get_graph_runtime",
             AsyncMock(
-                return_value=ActiveGraphRuntime(
+                return_value=GraphRuntime(
                     client=client,
                     entity_manager=MockEntityManager(),
                     relationship_manager=MockRelationshipManager(),
@@ -227,7 +227,7 @@ async def mock_graph_disconnected() -> AsyncGenerator[MockGraphClient]:
         patch(
             "sibyl_core.tools.search.get_graph_runtime",
             AsyncMock(
-                return_value=ActiveGraphRuntime(
+                return_value=GraphRuntime(
                     client=client,
                     entity_manager=MockEntityManager(),
                     relationship_manager=MockRelationshipManager(),
@@ -237,7 +237,7 @@ async def mock_graph_disconnected() -> AsyncGenerator[MockGraphClient]:
         patch(
             "sibyl_core.tools.explore.get_graph_runtime",
             AsyncMock(
-                return_value=ActiveGraphRuntime(
+                return_value=GraphRuntime(
                     client=client,
                     entity_manager=MockEntityManager(),
                     relationship_manager=MockRelationshipManager(),
@@ -247,7 +247,7 @@ async def mock_graph_disconnected() -> AsyncGenerator[MockGraphClient]:
         patch(
             "sibyl_core.tools.add.get_graph_runtime",
             AsyncMock(
-                return_value=ActiveGraphRuntime(
+                return_value=GraphRuntime(
                     client=client,
                     entity_manager=MockEntityManager(),
                     relationship_manager=MockRelationshipManager(),
@@ -258,7 +258,7 @@ async def mock_graph_disconnected() -> AsyncGenerator[MockGraphClient]:
         patch(
             "sibyl_core.tools.manage.get_graph_runtime",
             AsyncMock(
-                return_value=ActiveGraphRuntime(
+                return_value=GraphRuntime(
                     client=client,
                     entity_manager=MockEntityManager(),
                     relationship_manager=MockRelationshipManager(),
@@ -269,7 +269,7 @@ async def mock_graph_disconnected() -> AsyncGenerator[MockGraphClient]:
         patch(
             "sibyl_core.tools.health.get_graph_runtime",
             AsyncMock(
-                return_value=ActiveGraphRuntime(
+                return_value=GraphRuntime(
                     client=client,
                     entity_manager=MockEntityManager(),
                     relationship_manager=MockRelationshipManager(),
@@ -281,7 +281,7 @@ async def mock_graph_disconnected() -> AsyncGenerator[MockGraphClient]:
         patch(
             "sibyl_core.tools.admin.get_graph_runtime",
             AsyncMock(
-                return_value=ActiveGraphRuntime(
+                return_value=GraphRuntime(
                     client=client,
                     entity_manager=MockEntityManager(),
                     relationship_manager=MockRelationshipManager(),
