+++ packages/python/sibyl-core/src/sibyl_core/migrate/scope_backfill.py
@@ -50,17 +50,17 @@ import structlog
 
 from sibyl_core.backends.surreal.records import normalize_records, raise_on_error
 from sibyl_core.models.memory_scope import MemoryScope
-from sibyl_core.services.graph import (
+from sibyl_core.services.graph_entity_store import (
     _ENTITY_BULK_UPSERT_QUERY,
     CLEAR_MEMORY_SCOPE,
-    _entity_from_row,
     _entity_record,
     heal_entity_metadata_snapshots,
 )
+from sibyl_core.services.graph_records import _entity_from_row
 
 if TYPE_CHECKING:
     from sibyl_core.models.entities import Entity
-    from sibyl_core.services.graph import SurrealGraphClient
+    from sibyl_core.services.graph_client import SurrealGraphClient
 
 log = structlog.get_logger()
