+++ packages/python/sibyl-core/src/sibyl_core/migrate/verify.py
@@ -10,7 +10,7 @@ from sibyl_core.migrate.archive import (
     graph_payload_from_archive,
     validate_archive,
 )
-from sibyl_core.services.graph import normalize_records
+from sibyl_core.services.graph_common import normalize_graph_records as normalize_records
 from sibyl_core.services.graph_runtime import get_graph_runtime
 from sibyl_core.tools.admin import create_backup
