+++ packages/python/sibyl-core/tests/test_memory_spans.py
@@ -67,6 +67,6 @@ def test_snapshot_shadowed_keys_are_exactly_the_server_owned_ones() -> None:
     back from the snapshot after being withdrawn.
     """
     from sibyl_core.memory_pipeline.structure import STRUCTURE_METADATA_KEYS
-    from sibyl_core.services.graph import _SNAPSHOT_SHADOWED_METADATA_KEYS
+    from sibyl_core.services.graph_records import _SNAPSHOT_SHADOWED_METADATA_KEYS
 
     assert _SNAPSHOT_SHADOWED_METADATA_KEYS == STRUCTURE_METADATA_KEYS
